requirements.txt
Flask
flask-login
Flask-Mail
werkzeug
gunicorn
itsdangerous
pytest
psycopg2-binary
torch>=2.6.0
torchvision>=0.21.0
opencv-python>=4.10.0
Pillow>=11.0.0
numpy>=2.1.3
ultralytics>=8.3.33
